[FlexNG] Handling break rules
Adds break rule support for single-line column flex items (other
container types will be done in future CLs). This includes handling
break-inside/break-before/break-after rules.

A new LayoutResult status of kNeedsRelayoutWithNoChildScrollbarChanges
was added to allow us to handle a relayout for child scrollbars, as
well as relayouts as a result of fragmentation.

Bug: 660611
Change-Id: I23a3a2cdb2d471090d5d567a17b656e531965aab