Don't reuse generated breaks when space sequences become non-leading
The NNInlineItemsBuilder class is the responsible of creating the
NGInlineItems list form the text nodes in the DOM tree. A special part
of this logic is the one that inserts control items after a leading
space sequence, in order to introduce forced-breaks in that position.

When these leading space sequences are not in the paragraph's first line
it means that they are preceded by a line-break element. Removing such
break element triggers a new layout, since the DOM tree has changed.

The NGInlineItemsBuilder logic tries to reuse the existing items when
processing a new text layout. However, if a line-break element is
removed, we should also remove any control item previously added to
introduce forced-breaks, as commented before. The current code doesn't
do that, leading to the buggy behavior described on the bug.

This CL introduces an additional step in the AppendTextReusing funciton
to detect these kind of "orphan" control items and ignore them when
building the new items list.

Bug: 1263567
Change-Id: Ie9ebadd4952efb491b983fba1cd2c6a7d7d1bf0b